Atende às sugestões da review e protege rotas
